requirements.txt
anyio==3.7.1
aspose-words==23.9.0
bardapi==0.1.29
beautifulsoup4==4.12.2
blinker==1.6.2
browser-cookie3==0.19.1
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
deep-translator==1.11.4
docx2txt==0.8
exceptiongroup==1.1.2
Flask==2.3.2
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.48.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.3
Pillow==10.0.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
python-docx==0.8.11
pywin32==306
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
urllib3==1.26.16
Werkzeug==2.3.6
